SKETCH-1743: Make real structures from Sketcher not be queries
Some applications require that the output from the Sketcher
be a "real molecule", not a query. This takes the approach
that "query-ness" depends on the state of the whole molecule, and
follows from Hussein and Ricardo's earlier work.

In native RDKit:

* Reading from a SMILES means all atoms are Atoms: `assert not Chem.MolFromSmiles('*').GetAtomWithIdx(0).HasQuery()`
* Reading from a SMARTS means all atoms are QueryAtoms: `assert Chem.MolFromSmarts('C').GetAtomWithIdx(0).HasQuery()`
* When reading from SDF/molfile, individual atoms are typically `Atoms` unless they are `R` groups or have

The atoms in a molecule are Queries if any of the following are true:

* The atom is a Queryi
* The atom is an R-Group (these are also queries when RDKit reads from .sdf)
* The atom has a SMARTS string property
* The atom has any of the properties used to designate specific queries

We think that an "R group" is not _really_ a query, but it needs to
be a query for now so that it can be serialized correctly into
SMARTS.

This causes some changes in tests. The main one is that in some cases,
we get atom names instead of atomic numbers for SMARTS of elements that
have no aromatic form. For instance, Gold will be Au. You can see this
in native RDKit, too:

   Chem.MolToSmarts(Chem.MolFromSmiles('[#79]'))
   [Au]
   Chem.MolToSmarts(Chem.MolFromSmiles('[#1]'))
   [H]