SHARED-10045: Use C++ instead of Python HELM parser/writer
This removes usage of the python helm conversion api for coarse grain
mols and updates the converted ROMol to store a hash of the monomer id
as the atom's isotope number.
NOTE: Helm strings are still roundtripped through the python parser to
canonicalize the polymer information before converting to an ROMol.