BLDMGR-7627: Changes for mmshare to compile on Rocky Linux
Containerized "buildenv" was developed in BLDMGR-7514 and used in BLDMGR-7627 to build mmshare on Rocky 8. Minor updates to C/C++ to account for newer compiler on Rocky. Add linux distro checks to build system code in order for building to be compatible with both CentOS and Rocky, in the short-term. BLDMGR-7684 adds liblzma.so.5 to software/lib, required by Rocky environment.