SHARED-10518: Create a function to extract some specified atoms from a ROMol as a new ROMol
This would be useful to have. We're opting to create the new graph
instead of deleting unwanted components because the latter would be very
expensive for extracting a small number of atoms from very large
molecules. This is still faster for small mols (i.e. about 2x faster for small
mols and at least 3x faster for big mols( was 10x faster for
"CCC"*1000)).