SHARED-10955: structconvert hangs when trying to write a range of structures to a non-existent directory
When we open a file stream for reading or writing, we should check if the file
was successfully opened because the default behavior of std::ifstream and
std::ofstream is to not throw an exception, but to set the failbit/badbit state.
This means that an invalid file stream can still be written to/ read from without
running into an exception, which is what caused the infinite hang.

Ideally, we should always throw some type of exception when we encounter some
issue but that kind of behavior makes our existing logic more complex that it
has to be. I'm opting to just handle the cases where we have bad input arguments
as a way to address the reported issue. The steps to reproduce the bug
now result in a `RuntimeError: Couldn't open file for writing`