SHARED-10979: Molviewer WASM no longer seems to be able to read SDF!
I traced the problem to usage of the `file_path` parameter in the
`file_open_completed` callback when we import files. The `file_path`
parameter returns the base name not the actual path of the input file,
which caused the `get_compression_type` call to fail because it couldn't
located the input file.

The fix required me to add a way to determine the compression type of
the input file with just the string buffer, so I did that and refactored
relevant apis.