SHARED-11160: Cannot convert a helm string with large extended annotations to rdkit
The default lookup table results that attempts to minimize the size of
the output file results in hangs for very long inputs, so I'm switching
to a faster scanner. Adding a test in mmshare doesn't seem to make much
sense, so I'll add a performance test after this is merged

The new scanner alone improves the parse time for the provided example from
40+mins to 8s, but I figured it wouldn't hurt to just exclude the extended
annotations from the Bison/flex input since we don't need complex logic to
parse it. Doing that reduced the parse time by 10x to 0.8s without loss
in functionality.