SHARED-11549: Support #() in Custom Monomer Database
This updates the HELM parser to allow all non-whitespace characters
within a monomer id and to only recognize a monomer token as an inline
SMILES token if it is a valid SMILES token or contains whitespace
characters.

I tested the updated implementation with all monomer symbols and HELM
examples from the CycPeptMPDB database.