SKETCH-2091 add option to copy as image (#50)
* Linked Case: SKETCH-2091
* Branch: 2026-1
 
### Description
This adds the option to copy images and reactions as an image. It is
active from both the selection context menu and the general background
context menu (Copy All As...)

Full Reactions are rendered if a reaction is present and
-we're using the copy all as 
-we are exporting a selection, and the selection includes at least one
non molecular object.

If the user selects only molecules on a reaction, we export just the
selection as a molecule (the idea is that someone wants to copy a
product or reactant), but if the selection includes non molecular
objects we assume they mean the whole reaction (since we can't export
parts of the reaction) and give them that

### Testing Done
tested by copying as image and saving into a google doc

---------

Co-authored-by: Chris Von Bargen <christopher.vonbargen@schrodinger.com>
Co-authored-by: Rachel Walker <39069546+rachelnwalker@users.noreply.github.com>
Co-authored-by: Hussein Faara <hussein.faara@schrodinger.com>
Co-authored-by: Ricardo Rodriguez <ricrogz@users.noreply.github.com>