SKETCH-2540: Adding SketcherWidget::activateSelectionTool (#76)
* Linked Case: SKETCH-2540
* Branch: main
 
### Description
This adds a SketcherWidget method for activating any of the selection
tools. The only minor caveat here is that doing something like
```py
sk = sketcher.SketcherWidget()
sk.activateSelectionTool(sketcher.SelectionTool.ELLIPSE)
sk.setSelectOnlyMode(True)
```
won't actually activate the ellipse selection tool, since "normal"
Sketcher (i.e. not select only Sketcher) doesn't allow selection tools
to be activated when the workspace is empty, meaning that the
`activateSelectionTool()` will be a no-op. The workaround is to make
sure that select-only mode is enabled first, i.e.,
```py
sk = sketcher.SketcherWidget()
sk.setSelectOnlyMode(True)
sk.activateSelectionTool(sketcher.SelectionTool.ELLIPSE)
```
will work as expected. This seems like a relatively minor inconvenience,
especially since it's not an issue when using the (Python-only)
`SelectOnlySketcherWidget` class, as that will be in select-only mode as
soon as its instantiated.

I've also added SIP wrappings for the new method, but that part is
mmshare-only so it's not in this diff (since we don't have the SIP
wrappings in this repo yet). It's just the new line from
sketcher_widget.h copied over to _sketcher.sip.

@matvey83, GitHub still won't let me add you as a reviewer on this repo,
so I'm trying a ping instead.

### Testing Done
Tested with an mmshare build using the Maestro Python terminal.