LDID8-1005: publish @schrodinger/sketcher npm package during nightly build (#91)
* Linked Case: LDID8-1005
* Branch: main
 
### Description
Publishes an npm package with the WASM artifacts with a pre-release
version (so it doesn't need a version bump, the npm-publish action can
generate unique pre-release version names from the same base package
version as long as the commits are different). For now, will only
publish packages when run as a `workflow_call` (which currently only
happens from the nightly build workflow).

Merging it into `main` because it's practically not going to have any
effect on the `2025-4` branch (since the workflow file from that branch
will never be read in any of the cases where this action will run).