SKETCH-2550: Mouse cursor changes (#87)
* Linked Case: SKETCH-2550
* Branch: 2025-4
 
### Description
This PR makes a handful of changes to the mouse cursor to help it feel
more precise:

- The cursor outline, which provides contrast in case the cursor if over
something that's the same color as the arrow, is now dramatically
thinner. That way, even if the user is looking at the tip of the outline
instead of the tip of the arrow, the difference won't be substantial.
(The difference is now ~0.1 pixels, down from ~5.4 pixels, as a result
of both this change and the next one.)
- The cursor outline now uses a beveled join (via SVG styling), which
means that the tip of any outline corners are "chopped off."
- The cursor outline is now the exact same color as the background,
instead of a light gray.
- The cursor colors are inverted when Sketcher is in dark mode. This
will hopefully make the cursor outline less noticeable, since it'll
blend into the background, while the arrow itself will always contrast
with the background color.
- The cursor is now scaled down by 20% for Mac desktop builds, since
Mac's normally have a smaller cursor size than Windows. I haven't
actually tested whether 20% is the right value to use since I don't have
a Mac. (Although I _have_ tested the 20% scaling on Windows just to make
sure that it doesn't look ridiculous.) Would someone be able to test
this and see if the cursor size inside the Sketcher window is roughly
the same as the cursor size outside of it?

Here's a comparison of the old and new cursors. Note that I've added a
gray background for the new cursors, since otherwise the outline
wouldn't be visible against the background.

| Old | New (light mode) | New (dark mode) |
| --- | --- | --- |
| <img width="198" height="286" alt="cursor-arrow - orig"
src="https://github.com/user-attachments/assets/8213c3aa-8c7b-4b0f-88f4-45dacdb185e5"
/> | <img width="198" height="286" alt="cursor-arrow"
src="https://github.com/user-attachments/assets/e29b474a-2c53-43e6-82c7-ff0eb342c041"
/> | <img width="198" height="286" alt="cursor-arrow - dark"
src="https://github.com/user-attachments/assets/c318a557-0b94-47ef-9a73-c5eb1eb1a0eb"
/> |

### Testing Done
Tested this on Windows using both the normal cursor size and the
scaled-down-by-20%-for-Mac cursor size.