SKETCH-2563: fix keyboard shortcuts not working in the WASM sketcher (#101)
* Linked Case: SKETCH-2563
* Branch: 2025-4
 
### Description
The WASM sketcher was not responding correctly to any keyboard shortcuts
(except paste). This was because of a `mouseover` event handler in the
HTML responsible for focusing the sketcher on hover. That handler was
focusing the sketcher's shadow root's host element instead of the
sketcher's canvas inside that shadow root which messed up the event
propagation/handling in WASM preventing any of our keypress handlers
from being called.

### Testing Done
Manually tested in the browser in chrome and firefox