SKETCH-2569: Closing our export dialog before we open the OS dialog so that the OS dialog isn't immediately closed on Windows
* Linked Case: SKETCH-2569
* Branch: 2025-4

Description

The issue here is that FileExportDialog::exportFile() calls QFileDialog::saveFileContent() and then immediately calls accept(). QFileDialog::saveFileContent() launches an OS file dialog, but it does so asynchronously and returns immediately. The accept() call closes the FileExportDialog, and that dialog closure appears to be passed on to the OS file dialog. That seems a bit strange, especially because the OS dialog isn't parented to the FileExportDialog, but it's the only explanation I can come up with. Since it's a native dialog that doesn't know about Qt, maybe it natively "parents" to whatever window was active in whatever process spawned the dialog?

The issue was initially introduced by the Qt patch for SKETCH-2223. The change in modality must have affected whether the dialog closure gets passed to the OS file dialog. This PR fixes the issue by simply swapping the order of the two calls so that the OS dialog doesn't yet exist when accept() is called.

Testing Done

Tested the export dialog and the save image dialog on both Windows and WASM.