SKETCH-2565: fix tooltips on WASM with Qt::WA_AlwaysShowToolTips (#106)
* Linked Case: SKETCH-2565
* Branch: main
 
### Description
It appears that we're running into
[QTBUG-94583](https://bugreports.qt.io/browse/QTBUG-94583), which has
been around since at least Qt 5.15. One of the comments on the Qt case
mentions that setting `Qt::WA_AlwaysShowToolTips` on the window fixes
the problem. This attribute normally controls whether tool tips are
shown on inactive windows; e.g., if you had the desktop Sketcher open
but clicked over to a terminal window, would you get tooltips when you
hovered over a button on the inactive Sketcher window. For some reason,
this fixes the WASM bug even though the Sketcher window doesn't become
inactive.

Is this something that should go into 2025-4? I wouldn't _expect_ any
side-effects, but with Qt and WASM, who knows? At the very least, we can
be sure that the fix won't affect desktop builds.

### Testing Done
Tested a WASM build under Firefox and Chrome. If anyone wanted to test
this on their machine and the PR build isn't done yet, you can grab the
WASM build from
[here](https://github.com/KevKeating/sketcher/actions/runs/18409714866).