Only run test collection if the build succeeded (#112)
* Branch: main
 
### Description
Currently, the GitHub runner is set to run the test collection
`always()`, which means that it'll run even if the build fails. In that
scenario, the tests won't have been run, so test collection will fail
and clutter up the list of errors in the GitHub runner summary. This
diff changes things so that test collection is only run if the build
succeeded, so we know that there will be something to collect.

### Testing Done
Ran an automated build with an intentional configuration error and
confirmed that there were no errors from test collection.