Confirming that QFile::open() succeeds (#111)
* Branch: main
 
### Description
Qt asked us to test
SKETCH-2565/[QTBUG-94583](https://bugreports.qt.io/browse/QTBUG-94583)
with Qt 6.9 or 6.10. I tried to get a WASM build of Sketcher using Qt
6.10 to do the test, but the interface to Qt's WASM stuff has
significantly changed (yet again) so I gave up fairly quickly. As part
of this process, I updated our code for two minor non-WASM-related
changes in Qt 6.10.

- The `QCheckBox::stateChanged(int)` has been deprecated and replaced by
`QCheckBox::checkStateChanged(Qt::CheckState)`
- `QFile::open()` returns a Boolean indicating whether the open
succeeded, and that return value is marked as `[[nodiscard]]` in Qt
6.10.

For the second bullet point, I figured it's worth pushing those changes
to `main`, since the `QFile` check is useful and we'll need it
eventually.

It's not worth doing anything about the the first bullet point for now,
since the new signal doesn't exist in Qt 6.5, which is the version that
the core suite uses.

### Testing Done
Confirmed that the changes build and existing tests pass.