Cache dependencies as soon as their build succeeds (#110)
* Branch: main
 
### Description
Currently, dependencies are only cached if the entire build succeeds. If
the build of Sketcher itself fails, all of the dependency builds will be
thrown out. This changes things so that each dependency is cached as
soon as its build succeeds, which means that the cache will still be
populated even if there's a downstream failure. This can save quite a
bit of build time for the next build, especially on forks of the repo
where there's no nightly build running and so the cache isn't always
up-to-date.

### Testing Done
Ran a build.