SKETCH-2578: Removing the white line when using select-only mode and dark mode together (#119)
* Linked Case: SKETCH-2578
* Branch: main
 
### Description
The SketcherWidget has a line between the Sketcher workspace and the top
toolbar that stands out when the toolbars are hidden (i.e. select-only
mode) and a dark color scheme is used. This PR hides the line whenever
the toolbars are hidden, which fixes the issue.

### Testing Done
Tested using Matvey's test script.