SHARED-11884: Refactor HELM validation code to remove bison HelmParser api dependency (#141)
* Linked Case: SHARED-11884

### Description
<!--Provide a detailed description what these changes are and why they
are necessary -->
This allows us to replaces the bison parser. The changes don't add any
new logic, I just updated the validation apis to directly update the
list of validation errors instead of delegating to the HelmParser
instance. Additionally removed the use of some lambdas to make the code
more readable.