Add function to mutate a monomer in a monomermol (#133)
### Description
Adds a function that takes a monomermol and mutates a monomer at a given
index to a monomer with the provided symbol. Assumes that you are
mutating to a monomer of the same polymer type (PEPTIDE, CHEM, RNA).

This implementation does not currently throw if the linkages are
wacky/wrong/incorrect (see new tests) -- that will be added in the
future.

### Testing Done
See new tests. Tested that python wrapping works (although wrapping not
done in this repo).

Co-authored-by: Chris Von Bargen <christopher.vonbargen@schrodinger.com>