Fixing missing includes (that fail with fmt 12.1) (#145)
* Linked Case: NO_JIRA_ID

### Description
When I was trying to figure out some WASM issues, I tried upgrading to
the newest version of emscripten (4.0.18). Getting emscripten to work
required upgrading fmt to the newest version (12.1.0), since older
versions of fmt cause a build error when being built with the newest
version of LLVM (which is what emscripten uses under the hood). Getting
_that_ upgrade to work required adding these three includes to our code.
Apparently, we were relying on a transitive include for `fmt::join()`,
and fmt 12.1.0 no longer includes the transitive include.

I'm not sure if we want to upgrade `main` to the newest version of
emscripten right now. I've got the WASM builds working on my branch, so
it wouldn't be difficult to do. (It requires: this commit, bumping the
version numbers in versions.json, and one minor tweak to the CMakeFile
to ignore an overzealous warning.) Does anyone have any opinions on
that? A major upgrade of Qt (e.g. to Qt 6.10) would require the newer
version of emscripten, but I doubt we'd have a reason to do that any
time soon. Our current version of emscripten (3.1.51) is almost two
years old, and it looks like development work has been fairly regular,
although I have no idea if/how any of the changes would actually affect
us. There's nothing particularly significant about the jump from 3 to 4:
the ChangeLog says that going from 3.1.74 to 4.0.0 was just a normal
release.

Either way, we'll presumably want to upgrade emscripten and/or fmt
and/or Qt some day, so I figured I'd merge in the new includes while I
remembered them.

### Testing Done
Build succeeds.