SKETCH-2470: Implementing side bar and SketcherModel changes for monomeric Sketcher (#125)
* Linked Case: SKETCH-2470
* Branch: main
 
### Description
This PR starts implementing the interface for monomeric Sketcher. The
only things implemented so far are the side bar changes and the
SketcherModel changes, as well as the logic to connect those two things.
I haven't implemented any scene tools (those will be a separate PR), so
you can't actually draw a monomer in the Sketcher workspace yet.

By default, Sketcher still launches in atomistic mode, so none of the
changes here will be visible. To turn on the monomeric interface from
the command line, `sketcher_app` can be launched with an
`--allow-monomeric` flag. When using an embedded Sketcher, the
`SketcherWidget` ctor now takes an optional second argument that can be
`InterfaceType::ATOMISTIC` (the default), `InterfaceType::MONOMERIC`, or
`InterfaceType::ATOMISTIC_OR_MONOMERIC`. I'll put up a separate mmshare
PR that adds the monomeric interface to Maestro's Sketcher panel if a
feature flag is set, and that adds the InterfaceType stuff to the SIP
wrappings.

When reviewing this, it might be easiest to start with
`sketcher_model.h`, since that has docstrings for all of the new
`ModelKey` values, which gives a bit of an overall picture of how the
new logic is laid out.

Here's are some screen shots of the new changes. This is the atomistic
interface (i.e. all of the current tools):
<img width="851" height="572" alt="image"
src="https://github.com/user-attachments/assets/7877d388-9beb-47ba-9a65-c647c7104a8b"
/>

The amino acid tools
<img width="851" height="572" alt="image"
src="https://github.com/user-attachments/assets/72bd1ab6-35f7-4a6a-bfcc-f2989ba89593"
/>

The nucleic acid tools. The R(A)P and dR(A)P buttons allow the user to
draw a full nucleotide of RNA or DNA (which consists of three separate
monomers for the sugar, phosphate, and base). The other buttons only
draw a single monomer.
<img width="851" height="572" alt="image"
src="https://github.com/user-attachments/assets/4a95364f-16c3-442e-8293-3c8283bd116a"
/>

The RNA nucleotide popup. This allows the user to select a base for
their full RNA nucleotide.
<img width="851" height="572" alt="image"
src="https://github.com/user-attachments/assets/486799de-67e1-49ed-9300-03c44cf0568e"
/>

The DNA nucleotide popup. Both the RNA and DNA popups are implemented
using the new `NucleotidePopup` class.
<img width="851" height="572" alt="image"
src="https://github.com/user-attachments/assets/284baa88-1ff6-4cf8-b7a5-7f5cbb3cc788"
/>

There are a few changes here that look a little tangential to the case,
so here's a bit of context for those:

- The new monomer buttons are styled using `ATOM_ELEMENT_STYLE`, so I've
renamed that constant to `ATOM_ELEMENT_OR_MONOMER_STYLE`.
- `DrawToolsWidget` (which contains all of the atomistic draw tools) has
a "DRAW" label that changes to "EDIT ACTIONS" when there's a selection
in the Sketcher workspace. That label also needs to be visible when the
monomeric tools are active, so I've moved it to the `SketcherSideBar`
class, which contains a `DrawToolsWidget` as well as the new
`MonomerToolWidget`.

There's still a fair bit of work remaining for this case. All of these
things will be in separate PRs:

- The custom nucleotide button. This draws a full nucleic acid residue
(which consists of separate sugar, base, and phosphate monomers) and
lets the user specify any names they want for the sugar, base, and
phosphate. The SketcherWidget changes for this tool are implemented, but
not the button or the popup.
- The scene tools that handle actually drawing the monomers
- The keyboard shortcuts for the new tools
- The monomeric connector tool (i.e. the bond tool for monomers). This
has a separate case: SKETCH-2483

### Testing Done
Tested `sketcher_app` both with and without the new `--allow-monomeric`
flag.