Including versions.json in the dependency cache key (#153)
### Description
If we change the version of one dependency, we should trigger a rebuild
of _all_ of our dependencies, since some of our dependencies depend on
other dependencies. (For example, RDKit uses Boost.) This change adds
the hash of versions.json to the cache key, so any change in that file
will invalidate all cached dependencies.

### Testing Done
Ran a build on a runner on my fork.