SKETCH-2470: implementing monomer keyboard shortcuts (#144)
* Linked Case: SKETCH-2470

### Description
This implements keyboard shortcuts for the amino acid and nucleic acid
tools. The amino acid shortcuts are straightforward: hit a one letter
code to switch to that amino acid. The nucleic acid shortcuts are a
little more involved since there are four tools that could potentially
correspond to any given base type: the full RNA nucleotide tool, the
full DNA nucleotide tool, the full custom nucleotide tool, and the
monomer tool (which would draw only the monomer for the base, but no
monomers for the sugar or phosphate). Because of this, the keyboard
shortcuts works slightly differently depending on what tool is currently
active. Basically, if a full nucleotide tool is active, the keyboard
shortcuts will update the full nucleotide tool. If a monomer tool is
active, the keyboard shortcuts will switch to a different monomer tool.
Does this seem reasonable to people? Also, do the comments/code do a
reasonable job of explaining what key is supposed to do what/when/why?

### Testing Done
Tested using `sketcher_app` on Windows.