SKETCH-2619: Freeing up space on the memtest runner to avoid "No space left on device" failures (#155)
* Linked Case: SKETCH-2619

### Description
If the memtest runner has the build the dependencies (as opposed to
pulling them out of the cache), it winds up failing at the Build
Sketcher step with
```
/usr/bin/ld: final link failed: No space left on device
collect2: error: ld returned 1 exit status
```
Based on a quick Google, it's fairly common for the standard runners to
run out of space since they only start out with ~20 GB free. The
standard fix is to erase some of the preinstalled stuff that's not
typically needed. There are preexisting, configurable GitHub Actions
that try to do this for you, but I found this [blog
post](https://carlosbecker.com/posts/github-actions-disk-space/) that
says he got most of the same benefit with an `rm` call and some Docker
image pruning, so I tested that to start with. The Docker image pruning
didn't appear to do anything for us (maybe because we don't build Docker
images are part of our build process?), but the `rm` call nearly doubled
our free space. It deletes:

- .Net libraries
- Android libraries
- a Haskell compiler
- a security-focused static code analysis tool

and we don't use any of those. I added some output to the cleanup step
so we could see the effect, and got
```
Free space before cleanup = 19G
Free space after cleanup = 37G
```
That was enough additional free space for the memtest build to pass on
my fork, even when it had to build all of the dependencies. The step
takes about a minute so I currently have it set to only run for the
memtest build, but we can enable it on the other ubuntu-latest runners
if we ever start running out of space on those builds.

### Testing Done
Ran a build on my fork of the repo where none of the memtest
dependencies were cached.