Sketch-2534 avoid zooming out in select-only mode (#150)
* Linked Case: <SKETCH-2534>

### Description
I moved the existing variable that keeps tracks of the activation of
select-only mode from the view to the model, so it can be accessed by
`Scene`. When a fit to screen is required the zoom out is now optional
and is skipped if we are in select only mode.

### Testing Done
tested in molviewer. Existing tests pass