SKETCH-2481: Prevent the user from mixing atoms and monomers when pasting or importing (#160)
* Linked Case: SKETCH-2481

### Description
This branch prevents the user from pasting or importing a monomeric
model if the SketcherWidget is atomistic-only or if the Sketcher
workspace already contains atoms, and it prevents the user from pasting
or importing an atomistic model if the SektcherWidget is monomeric-only
or if the Sketcher workspace already contains monomers. To implement
this, I had to do a bit of refactoring to `add_text_to_mol_model`. It's
now split into two different functions:
`convert_text_to_mol_or_reaction` and
`add_mol_or_reaction_to_mol_model`. (`add_text_to_mol_model` still
exists; it's just a two liner that calls the new functions.) This way,
the `SketcherWidget` can call the two new functions directly, and make
sure that the molecule is the right type (atomistic or monomeric) before
it gets added to the MolModel.

### Testing Done
Tested in standalone Sketcher on Windows and added a unit test.