SKETCH-2632: Allow copy of partial selection if no reaction elements selected (#173)
* Linked Case: SKETCH-2632

### Description
We somehow lost this behavior in the move to molviewer -- when a partial
reaction is selected, so long as no reaction elements are selected, we
used to still export as molecular formats. Looks like some LD customer
has been doing this for a long time and wants it back; seems like a
reasonable thing to restore.

### Testing Done
Added a test; CI