SKETCH-2617: Very basic functionality for the monomer scene tool (#172)
* Linked Case: SKETCH-2617

### Description
This is the start of the monomer scene tool implementation. It's missing
a ton of functionality; so far all it can do is display a placeholder
cursor hint and add a new disconnected monomer when clicking on empty
space. I figured I'd put up a PR while the line count was still
reasonable, though, since this at least conveys the overall arrangement
of how things fit together. The main changes here are:

- I added a `MolModel::addMonomer` method to actually add the monomer.
This uses `rdkit_extensions` to create the new monomer's `RDKit::Atom`.
- As part of that, I exposed the `rdkit_exensions::makeMonomer`
function, which creates a new atom without adding it to any molecule
(since MolModel already has logic to undoably add a new Atom)
- I created the new scene tool class, following a vaguely similar-ish
organization to the scene tools for adding atoms.

I also fixed two minor bugs that I found while testing out the above
changes:
- Previously, it was impossible to switch from the select tool directly
to a monomer tool, since enabling the monomer tool failed to set the
DRAW_TOOL sketcher model key, so clicking on the monomer button wouldn't
deactivate the select tool.
- The predictive and selection highlighting for nucleic acid bases
(which use the diamond shape) were noticeably smaller than the
highlighting for the other monomers since we didn't account for the 45°
rotation of the diamond.

### Testing Done
Ran Sketcher in standalone mode and confirmed that I can add monomers.