SKETCH-2635: Fixing side bar spacing in WASM builds (#181)
* Linked Case: SKETCH-2635

### Description
The side bar uses a QStackedWidget to switch between the atomistic tools
and the monomeric tools, but one consequence of this is that the both
sets of tools are allocated as much vertical space as the larger of the
two. The monomeric tools have a vertical spacer at the bottom that
"absorbs" the extra space, but the atomistic tools don't have anything
equivalent. This isn't an issue on the desktop builds, but apparently
the spacing is slightly different on WASM builds. As a result, the
R-group tools were being given extra space, resulting in a gap above the
buttons. This PR adds a vertical spacer to the atomistic tools, which
solves the issue.

### Testing Done
Built a WASM build on my fork and confirmed that this fixes the issue on
Firefox on Windows.