SHARED-11988: Move sketcher-only code out of rdkit_extensions (#188)
* Linked Case: SHARED-11988

### Description
Rachel and I are on a campaign to start thinning out rdkit_extensions --
the first easiest pass is to move all sketcher-specific code out. I'll
explore moving any of this into RDKit itself at a future point.

The only interesting thing was the collision between the various
find_centroid functions -- collapsing them uncovered an amount of
redundancy, which is resolved in the review.

### Testing Done
CI