SKETCH-2496: Remove dry-run; add ability to manually dispatch an NB (#191)
* Linked Case: SKETCH-2496

### Description
I updated bump-and-merge so that it is skipped if a token is not
supplied (the workflow would fail otherwise!) which allows us to
manually trigger an NB run.

Also, @lennax had added a dry-run option to the NB workflow so she could
test triggering it from core-infra-actions. That all works now, and the
dry-run options doesn't actually *do* anything (the bump, merge, and
build steps don't have dry-run options). Removing it unless folks think
there's utility there.